requirements.txt
pyowm~=2.10.0
pyTelegramBotAPI == 3.6.7
python-dotenv == 0.13.0